+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Default value for "nativetheme" to "dark"
 - Default value for "tocpages" to "vieb://help"
 - Disabled favicons to be empty instead of showing the broken image indicator
+- Mapsuggest and notification alignment in default colorscheme
 
 ## Removed